Use origin namespace instead of federation prefix for token verification
When the broker code was initially written, there was only one prefix for
the origin, so all of the code was written assuming a one-to-one mapping
between the origin and the prefix. This is the first step in breaking that
assumption by using the origin issuer key for token verification instead
of the key associated with the prefix.